Enable "set -e" by default
Run stack.sh with "set -e" to make devstack exit immediately if something
fails. This should make debugging easier as we will exit on the first
sign of trouble, so it will be easier to find the root cause.

Change-Id: I39e2c07916e0440c6bd1b632eaf792b2fa9e4bb5
WIP: I assume we want to make this configurable like  we do for nounset