Update patch set 7
Patch Set 7:

There is too much round discussion on if we should make 'port-sec' as a default extension driver, and should we leave a choice to user to make extension driver empty, without 'port-sec'.

IMHO, ':-' would be OK, just like the other config item. If we leave a chance of using empty extension driver, the db schema of the network/port created could not be updated if the user manually add 'port-sec' some day later.

@rkukura what's your opinion?

Patch-set: 7
Label: Verified=0