Update patch set 2
Patch Set 2:

I am really not a fan of relying on PATH to find privileged executables.  I know this is DevStack and security is not at the top of out priority list, but PATH in sudo and rootwrap is just a bad practice and I'd rather not use it in a place where people will point to it and say "but DevStack does it".

IIRC rootwrap itself actually encourages this practice and I've called it out in a couple of places, maybe its time to submit the changes to get rid of the practice?

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0