Update patch set 2
Patch Set 2:

The goal here is to facilitate use of the neutron HA feature in terms of root wrap configuratoin. We have tried a few approaches so far:
  -- add local filter in l3.filters
  -- add /usr/local/bin to exec_dirs in neutron/etc/neutron/rootwrap.conf, which will then be copied to /etc/neutron/rootwrap.conf during ./stack.sh
  -- modify lib/neutron-legacy to add /usr/local/bin to exec_dirs  in /etc/neutron/rootwrap.conf 

I see that everyone has a good point preferring one to the others. Sean has a good point indicating that all the exes are installed in /usr/local/bin on Ubuntu/Debian. So the second approach seems to be reasonable.

Patch-set: 2