Wait after restarting rabbitmq
This change Ie45446d3817b2f15631f03b2af84749fe936c67b generalize
the restart of the rabbitmq service.
However, on some platforms, the service may not be ready when
devstack wants to use it.
This commit introduces a call to the "rabbitmqctl wait" command.
The default rabbitmq PID will be used if it exists.

Change-Id: I48bb54742d667f551e3c9413581dd677e832228b
Closes-bug: #1449056