Update patch set 2
Patch Set 2:

Thanks all. @Sean The reason is simple. I just found a failure from the VM start when I use a 32bit host, it's because nova try to boot a 64bit cirros.
I think it's easy to fix if we add the check for the host. yes, maybe 64bit is more popular, and 32bit host is unusual. So I made the default condition to 64bit cirros.
For the multihost setup, only the node which start g-req will try  download images and upload it to glance. And suppose that the compute nodes include the 32bit and 64bit, this fix could not improve it indeed, but the fix is not for that. even if we don't use this fix, there still be this kind of problem.

Patch-set: 2
Reviewer: Gerrit User 11114 <11114@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0