Fix IMAGE_URLS defaults for xenserver and ironic
Xenserver and Ironic were both unconditionally tacking on an additional image
to the `IMAGE_URLS` list. If the user specified that image themselves (as
directed by current documentation), then devstack would attempt to upload the
image to Glance twice, resulting in a 409 error, and devstack failing.

The fix is to only tack on that additional image if `IMAGE_URLS` has not been
supplied by the user.

Change-Id: I2eacf34f6361289abd82b824e82723034995ea84
Closes-Bug: 1451623