Cinder: Set os_privileged_user credentials (for os-assisted-snapshots)
When calling os-assisted-snapshots APIs, Cinder often (by default) needs to pass an admin token to Nova. Currently it uses the credentials of the current user.

This will cause calls to Nova APIs for assisted volume snapshots to fail.

Configuration options should be added to specify different credentials for talking to Nova.

Change-Id: I9e3ed53f4e1349d57a0c33518445f54ac63e36ec
Related-Bug: #1308736