Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

I like this, thanks for jumping in Keystone guys...

Any changes to IDENTITY_API_VERSION need to happen in stackrc as other files need a correct value from there.

I'd also prefer to see the variable named ENABLE_IDENTITY_V2 but that is not a blocker.  It would be more consistent with other settings of this type.

Patch-set: 6
Label: Code-Review=-1