Update patch set 8
Patch Set 8:

Ian, honestly I do not think moving the existing configuration is related to this patch.

Looking at the new code, it is clear that IDENTITY_API_VERSION will be set to 3 if ENABLE_IDENTIY_V2 is False, independently of its previous value.

The existing IDENTITY_API_VERSION definition sets the default value, which can also be overridden by local.conf

My point is to let this change atomic, as it is. Feel free to submit another patch to fix that, I just do not want to introduce unrelated changes.

This is what I think, let's see other people views.

Patch-set: 8