Update patch set 1
Patch Set 1: Code-Review-1

I too would rather see people adjust jobs to work with tox 2.0 rather than use an older release. If we do that, we don't really know what's broken and then it will be a much bigger production to undo the cap once we think we might be ready (which means nobody is going to want to be the one to take it on).

It's early in the cycle, now is a good time to solve this. If it were release crunch my opinion might be slightly different.

Patch-set: 1
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1