Update patch set 7
Patch Set 7: Code-Review+2

For now, all of the images that we use in the gate have to be listed here.  We can talk about moving them later, but we don't have a mechanism to do that with plugins yet.  In the mean time, this is now dead code, so it would be nice to get it out of devstack.

Patch-set: 7