Passing correct tenant/user vars to context log
Currently the deployment of devstack + Ironic is breaking due to
changes to comply with oslo_context and the log formatting string
that expects, by default, user_name and project_name in the
context. This patch aims to pass the current variables in the
context of Ironic to generate a correct log formatting string.

Change-Id: I44e4490909a117930b1883dac859e0213827c834
Closes-bug: 1456428