Update patch set 1
Patch Set 1:

Part of your problem is in assuming that rejoin actually works.  as you see, it often doesn't.  This really is not something that is a high priority as that script does not always start processes properly.  I am not interested in changing the process functions to accommodate this without a lot of testing to ensure that the rest of this insanely fragile setup doesn't break.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0