Update patch set 2
Patch Set 2:

Actually, I ran into this on a pre-existing setup which I've been running for about a year or so. Eventually the logs accumulated and were too much. I know of other people it happened to as well.

Can you please clarify why it is clumsy?

Patch-set: 2
Label: Verified=0