Update patch set 6
Patch Set 6: Code-Review+2

I'm fine with that paragraph Ian, it is still accurate, the change adds single interface setup, which still applies in a multi-node cluster.

However, if this does get re-spun, 'Neutron' needs to be capitalized in the title.  I want to do a general cleanup for the docs soon and will catch it then if not here.

Patch-set: 6
Label: Code-Review=+2