Use keystone v3 API for projects
Always use the keystone v3 API for project creation. Make domain a required
argument. Whilst we could simply default this value within the function I think
it's better to make this explicit as these are things deployers and services
need to consider.

In future we will want to figure out how we want devstack to organize domains
however I don't believe that it belongs in this patch.

Change-Id: Ib9587193c5c8419dc4b5a608246709baaddd2a52
Implements: keystonev3