Update patch set 4
Patch Set 4:

As I noted in the bug report, I think it is not necessarily an error of the restart script not to wait for the service being completely available. So maybe the approach I just proposed works better, only doing the restart after the initial attempt to setup the rabbitmq user fails.

Patch-set: 4