Revert "Configure boto.cfg in tempest to point to SSL bundle"
This reverts commit 90c20209088eca3dcf830856a3a3118367dbc09f.

The sed could fail due to permissions on creating a temporary
file in /etc, causing setup.sh to fail.

The boto tests are no longer executed in the tempest smoke
test, the reason this change was made. So rather than trying
to work around the permission issue via sudo, simply revert
the unnecessary patch.

Change-Id: I5f4c43bbbe477c570936e2e40ac05cc38febbb3f