Update patch set 7
Patch Set 7: Code-Review+2

WTF re 'Neutron' in the title?  It's the project name and the name is properly capitalized everywhere I see.  There also happens to be a 'neutron' command that should never be capitalized as it is a command literal string, but that's not the reference in the title.  Ugh, I'll go seek out examples of this and find out why the rest of the world is doing it wrong.

FWIW, this isn't the docs repo ;)

Patch-set: 7
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2