Update patch set 1
Patch Set 1:

The problem is that the sed in iniset is done in /etc, so while you may have write access to the file you don't have permission to create a temporary file in /etc.

I added this bit because the boto tests used to be included in the tox smoke test which is the test suite I'm targeting for devstack-gate for SSL/tls-proxy testing. Since boto was removed I took the easy way out and submitted this.

A couple of other possible solutions, and I'd be happy to examine them, include:

- Create a temp dir, copy /etc/boto.cfg there, iniset it, copy it back
- Add a sudo to the iniset command. Makes me a little queasy but there is a lot of other sudo

Patch-set: 1
Label: Verified=0