Ensure netfilter enabled for bridges with OVS Hybrid
The OVS Hybrid driver implements security groups by using netfilter on
an intermediate Linux bridge.  Unfortunately distributions differ on
the default status of this (Ubuntu has it on, Red Hat off, at least).

Neutron is unwilling to set this up itself
(If2d316eb8c422dc1e4f34b17a50b93dd72993a99) but I have a review out to
detect this situation and halt the agent
(I8b49298d27a75a864c0066896cbd3d3095982f6e).

This change enables the setting, and does some minimal work to
configure the system to have it enabled.  This is analgous to what is
done in other deployment mechanisms such as packstack and fuel, and
also what is suggested in the documentation.

Change-Id: Ia508c552d806363f6633d0bcd357c05bf956279f