Update patch set 5
Patch Set 5: Code-Review+1

Tempest is branchless and this feature has been integrated since Kilo.
So we need this kind of patches for both master and stable/kilo branches.
But we cannot set True as the default value of Tempest because stable/Juno has not the feature and we need to still test the branch now.

Patch-set: 5
Reviewer: Gerrit User 6167 <6167@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1