Start and stop n-net with n-cpu for partial upgrades
In a real environment, n-net will be co-located with n-cpu, and
upgraded at the same time. For the sake of expediency, this patch
makes n-cpu and n-net be started and stopped at the same time as
"the compute family of services." Grenade's existing partial upgrade
support will still apply, and we'll get a more realistic test.

Depends-On: I95b085fbb40ad07777cdac322908b6e6d31aa5b1
Change-Id: I7a2f3ee9c74b34fdf003f65f4936074a236784c2