Start and stop n-net with n-cpu for partial upgrades
In a real environment, n-net will be co-located with n-cpu, and
upgraded at the same time. For the sake of expediency, this patch
makes n-cpu and n-net be started and stopped at the same time as
"the compute family of services." Grenade's existing partial upgrade
support will still apply, and we'll get a more realistic test.

Since grenade uses devstack old to do the shutdown, we need to backport
this patch to Kilo to make the Kilo -> master (Liberty) job work.

Change-Id: I7a2f3ee9c74b34fdf003f65f4936074a236784c2
(cherry picked from commit 18e9151b03583ba2426e95fc74d1733633777eb0)