Update patch set 4
Patch Set 4:

Looks like maybe n-net isn't retrying its connection to the message bus, which causes nova-compute to fail to talk to it. Not sure why that'd be different, but I'll take a look.

We need to fix it, but in reality an admin or monitor would restart nova-network I think, so I wouldn't say this has any damning implications in reality. Just needs to be more robust.

Also, the e-r filter for bug 1315095 needs to be changed because it will always trigger in failed partial jobs now I think (since there will be new/n-net.log file).

Patch-set: 4