Add required devel packages for python-qpid-proton
python-qpid-proton does something similar to pyzmq when libzmq is not
installed. It bundles the underlying library and builds it.
Unfortunately, python-qpid-proton currently has a dependency on
`uuid-dev` and `swig` to build the python module.

These dependencies are being relaxed but it won't be until the next
release that we'll be able to live without them. Until then, I'm adding
these packages to the list of required packages.

Change-Id: Icb22b710d2a7eeaed6d8fa55a389077204d1f8d9