Update patch set 1
Patch Set 1: Code-Review-1

so, I know this is how we ghettoed things up in nova to do partial upgrade support, but that predates us having multinode facilities upstream.

I really think that all the partial upgrade work should be done at the multinode level, and just not upgrade the worker node. This is a much more realistic setup for the tests, and should actually make the grenade and devstack code simpler as well (as we won't need all these odd service conditionals).

Patch-set: 1