Add IPv6 support to openrc files
Added -6 argument to openrc to set the keystone URL to be an IPv6
address. Assumes devstack was configured with SERVICE_IP_VERSION=6 in
local.conf

Additionally, SERVICE_IP_VERSION is stored in .stackenv and checked in openrc. This means that the "-6" argument is optional and if SERVICE_IP_VERSION is set to 6, openrc will use IPv6.

Change-Id: I5d46d5438d3e56fea788720ca17f0010caef3df1