Update patch set 3
Patch Set 3: Code-Review-1

Over 95% is not 100%.  The qpid codepath especially is working, and without testing it, it will likely fall behind; this is without even considering up-and-coming 0mq, which will have an even harder time gaining adoption as a result.  oslo.messaging has made a decision of what to support; I don't think that devstack should go a different path, especially since this code is already written and working.

Patch-set: 3