Update patch set 3
Patch Set 3:

a number of the current integrated release projects have devstack support via external devstack plugins. I don't understand why it's a big deal to have non main path rpc servers do the same. This is the direction devstack is going in the big tent. 

oslo.messaging might be the right place to host some of these if it's decided on such broad support.

Patch-set: 3