Actually install the requirements repo.
The requirements repo has had a setup.cfg etc for a long time but only
recently started using it. As it now has dependencies, we need to pip
install it. To preserve compat with older requirements repos I haven't
changed the call to invoke update-requirements yet, as we still have
the update.py symlink.

Change-Id: I7d7e91694c9145fac0ddab8a9de5f789d723c641