Update patch set 3
Patch Set 3:

Folks, i respectfully disagree with the bit of over reaction happening here.

It's not that hard to do a devstack plugin for zeromq.

http://docs.openstack.org/developer/devstack/plugins.html

The email from Sean and this review was a nudge to folks interested to get that effort going instead it turned into this -1 fest! we are still a week before L-1. this was the right time for the signal and for getting the ball rolling...

-- dims

Patch-set: 3