Running with 'LIBS_FROM_GIT=pbr' causes failure
If you have 'pbr' in your LIBS_FROM_GIT list, then the following error
occurs:

+ cd /opt/stack/requirements
/home/ubuntu/devstack/inc/python: line 229: cd: /opt/stack/requirements:
No such file or directory

This is due to commit: https://review.openstack.org/#/c/192494 moving
the clone of 'requirements' after 'setup_lib "pbr"' - which depends
on the directory existing.

The fix is to move the clone back before the pbr section,
leaving the pip_install for afterwards.

Change-Id: I04bd887d3c8157533f79cb737d040a5803c7a9dd
Closes-Bug: #1466577