Update patch set 1
Patch Set 1:

Well , I do have a single support, but I think the OS needs to do what do do with the VMs coming on the interface in devstack ... I found this on a forum and I tried it and it worked.  Perhaps putting it in a function would be better.

Dave

Patch-set: 1
Label: Verified=0