Set compute-feature-enabled.attach_encrypted_volume=False for Ceph job
The encrypted Cinder volume tests in Tempest don't actually work
properly for a Ceph backend in cinder since the volume encryption
support is not in Nova for RBD volume types.

This disables the encrypted Cinder volume Tempest tests when the
storage_protocol is 'ceph', which is set in devstack-gate for the Ceph
job.

This is needed for Cinder change
I03f8cae05cc117e14f7482115de685fc9f3fa54a which tells Nova that the rbd
volume connection is on an encrypted volume type.

Depends-On: I48eba7c645cc1c979fd766ae9c05efb00957f787

Related-Bug: #1463525

Change-Id: I9e12f8dc9e3e6b68dc031351cb081ee2bc6e6cbb