Use path-mounted keystone when running in httpd
When running in httpd, keystone accepts requests on /identity and
/identity_v_2_admin (essentially deprecated given v3, also, note
that 'identity_v2' isn't used to avoid confusing applications that
look for version strings in the url).

The path endpoints should be preferred over the ports so the
server is configured to point applications to the path endpoints by
setting admin_endpoint and public_endpoint.

Change-Id: I34569b9e03c3f36748c92d803349e22a7ee1a633