Update patch set 3
Patch Set 3:

So this might look a little ugly. https://review.openstack.org/#/c/189250/ was breaking without this.

We can unwind this and discard it once we have constraints in place, but we can't get there without the requirements changes landing, which requires a modern pbr consistently available. Fixing all the deps will be done, but just *one* can break us again, even if its not one of our trees, so I think this is an essential belts-and-braces for the interim.

Patch-set: 3
Reviewer: Gerrit User 4190 <4190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0