Update patch set 1
Patch Set 1:

Yeah, I like the proposed architecture.  It's also more work and I really just want to "stop the bleeding" ASAP by getting to where we can block Neutron changes that break live upgrades.  So, part of me really doesn't want to block on reworking things to work as a multi-node test.

What about letting all of this in and working on the multi-node setup with less time pressure?

Patch-set: 1