Update patch set 3
Patch Set 3: Code-Review-1

I really feel like this is hitting things with a giant sledge hammer that is going to be hard to unwind, because the work around won't be able to be removed until a bunch of other things happen.

The right fix here is to get those other things happening first. It slows down the requirements work, but it doesn't put a kludge in that can't easily be removed later, and causes masking of real issues.

Having run into a patch that was rushed in for just such a kludge today that's masking real issues (https://review.openstack.org/#/c/194192/) I don't want to do any more of this.

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1