Update patch set 2
Patch Set 2: Code-Review+1

So what is happening here is that PIP_VIRTUAL_ENV is a var that pip will consume directly so any pip run in the outer function was using that virtualenv. By moving the setting of that var into an inner function we isolate that state for the pip runs in the outer function. LGTM.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1