DO NOT MERGE: nova-compute GDB backtrace
In trying to figure out what nova-compute is doing when it hangs in the
multinode job, we now trigger a guru meditation report to see what is
going on. But nova-compute isn't responding to that when it hangs, so
the next step is to get a gdb backtrace.

Hopefully we will see bug #1462305 on this patch soon so I don't need a
dozen rechecks

Duplicate of I89ba5f11de86c769e7db10cb1e1faa6e0a369aed to increase the
odds of hitting bug 1462305

Change-Id: Ib7fbf62981b4f48e949e4d1122f9fa17b7f31e4e
Related-Bug: #1462305