Update patch set 1
Patch Set 1: Code-Review+2

Sean, I'm OK with this, it moves in the direction that we're headed anyway with or without multinode.   I told Russell to not wait for the Neutron rewrite to do some of this.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2