Update patch set 2
Patch Set 2: Code-Review+1

This makes sense for devstack, though I am not a huge fan of referencing out of the tree directly (because someone is going to do it in production and complain when there is a security issue).

I see this as a reasonable compromise in the devstack land to avoid issues down the line.

Patch-set: 2
Label: Code-Review=+1