Update patch set 1
Patch Set 1:

I think that this patch should not be abandoned. Having this done in DevStack makes a lot of sense. Kyle's patch contains a plethora of reasons why having this as a default in the source tree is a bad idea, and Nova should stop doing it as a matter of fact, but that's a different story.

Patch-set: 1
Reviewer: Gerrit User 748 <748@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0