Update patch set 1
Patch Set 1:

All but one of the files being removed here are purely iniset calls, which can be done in local.conf directly, no plugin required.

The exception is the netapp nfs plugin and it simply adds a cleanup function that unmounts any mounted nfs volumes, which is only used by unstack.sh and/or clean.sh.

Patch-set: 1