Update patch set 1
Patch Set 1:

I am resurrecting https://review.openstack.org/#/c/190716.  In that review, you'll find a link to a change in neutron to set the default for all of neutron.  There are some privacy concerns for setting that globally, so here's another round of setting it in devstack, to whatever the local system is using.

Patch-set: 1
Label: Verified=0