Update patch set 1
Patch Set 1: Code-Review+1

I'm on board for this - Mark McClain is going to write a patch on the Neutron side to print a warning at startup if the setting is left unset, so that we can alert new deployers to this setting. There will also be some documentation put together.

LGTM

Patch-set: 1
Label: Code-Review=+1