Remove ceilometer in favor of plugin
The ceilometer project is moving to using a devstack plugin rather
than having ceilometer in the base devstack. This is to allow
greater control and flexibility.

At the moment this is somewhat hung up on trying to decide how to
deal with a project which was once integrated in both devstack and
grenade and now wants to be plugins in both. Those issues are being
worked out in multiple patches.

Depends-On: If7bb809743ab0c475a1844c79e552d8b6892a18f
Change-Id: I413ab159474b7d7231ad66d3a482201f74efe8a8