Update patch set 1
Patch Set 1: Workflow-1

I agree that an env var would be handy.  Replacing config files is more of a power user thing, and setting your DNS is a pretty basic setting.

This begs the question of whether support should also be added for nova-net, but I'm going to start with adjusting for content and style as suggested (n-net that might be a follow-on patch.)

Patch-set: 1
Label: Workflow=-1