Update patch set 3
Patch Set 3: Code-Review-1

a) Manually setting API_WORKERS=0 is an error, if anything we should abort on that condition in every case; it has not been a problem thus far.

b) If API_WORKERS gets auto-set to 0 (anything <2 actually) that is a bug in stackrc and needs to be fixed.

Patch-set: 3
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1