Update patch set 1
Patch Set 1:

Chris, I understand your thinking here, but if anything other than Swift (remember out-of-tree plugins) wants to use that it will be non-intuitive for them to need to remember to include lib/swift (if not already) to find the install function.  This is likely to show up in Grenade first and be hard to diagnose.

In other words, this breaks our pattern and that gives me hives. ;)

Patch-set: 1
Label: Verified=0