Update patch set 1
Patch Set 1:

Grr...so I submit that and see the follow-on patch...

We need to think through this because it is going to come up again I'm sure.  But I think the solution is to change how lib/swift works and make it detect if ceilometer is configured and/or that install function is present.  If we do the plugin right the function will be present by the time swift wants to call it.

Patch-set: 1