Update patch set 1
Patch Set 1:

Dean, have you seen this thread: http://lists.openstack.org/pipermail/openstack-dev/2015-June/068162.html ?

ceilometermiddleware has nothing to do with ceilometer other than its name. What the package contains is a middleware for counting swift requests, and only swift requests (it processes paths to extract container and object information).

Also, there's not requirement that ceilometer be the recipient of the notifications produced by the middleware, it can be anything that is willing to interpret the notifications on that topic. This decoupling is a good thing.

So this isn't, at least as I understand it, a question of discoverability but rather of "how should this be turned on and by whom".

I'm not sure what this means in terms of bring these changes to fruition, but am happy to make it happen based on whatever input anyone can provide.

Patch-set: 1