Update patch set 9
Patch Set 9: Code-Review-1

Just to summarize from our irc discussion on this. The confusion was because strict mode was actually removed in change 7af8a1b9b3180da54e2c9505228ad722db44ca27

After thinking it over a bit more I think we shouldn't use the old string as a selector behavior and definitely not default to strict. (since it doesn't exist anymore) Instead I think we should just make the new variable about the use of constraints (or not using constraints) which is True or False.

Patch-set: 9
Label: Code-Review=-1