Enable pipefail
Enable pipefail so we notice when intermediate commands fail.

This includes a couple of work-around for failing cases.

It is probably arguable that

 $ echo "foo" | grep " bar " | ... do something

is a false-positive of pipefail, if "do something" can handle a blank
string.  I think that the benefit of pipefail outweighs it and you can
write the same thing better in other ways.

Note some overalap with [1] which is where we first discovered these
issues.  Keeping this as a separate thing from the keystone v3 changes
though.

[1] https://review.openstack.org/#/c/186682

Change-Id: If2daad2f233f80dc274603a3f1dac2260bd00ea6