Update patch set 2
Patch Set 2: Code-Review+2

Chris, I had not seen that thread, thanks.  I did fall for the it-has-ceilometer-in-the-name-so-it-must-be-ceilometer bait.  We still have the issue of other projects wanting to install this needing to know to source lib/swift to get it, but until that happens (if ever) this is fine.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2