Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I do know at least one other 'not-devstack-thing' calls this script but I can not think of what it is just now.  Since this is in tools/* it is an implied interface and we should at least recognize --os-tenant-* for a time, but no longer document them.

Patch-set: 1
Label: Code-Review=-1