Update patch set 1
Patch Set 1:

I don't remember the specific details of the grenade problems, but they didn't actually appear until the predecessor of this patch (https://review.openstack.org/#/c/153079/) was committed and merged.  At that point all of the check and gate jobs broke, and it wasn't pretty.  The above change was reverted by https://review.openstack.org/#/c/160913/

Because of this past history, we do need to be very careful about merging this.

Patch-set: 1