Fix stack failure when default subnetpool is set
Currently stack.sh will fail if a value is set for
default_ipv4_subnet_pool and/or default_ipv6_subnet_pool in
neutron.conf. This is because setting either of these values
overrides the default behaviour of using the implicit (none)
subnetpool for subnet creation.

This patch fixes the failure by specifying subnetpool = None
in calls to subnet-create, so that neutron will behave as devstack
expects.

Change-Id: I29b2d62a022b43f6623b127af2ca303f9de847b0
Closes-Bug: #1472200