Update patch set 2
Patch Set 2:

It seems like the problem is restarting rabbit inside a loop...so move the restart outside and just wait for it to come up?

If there are differences between distros and distro releases on detecting if a service is running that should be wrapped up in a function and hidden.  Maybe using the distro-provided tools to do this rather than pgrep would also be helpful?

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0