Update patch set 4
Patch Set 4: Code-Review-1

@Doug and I were discussing this patch the other day. Honestly, I'd really like to keep things like this out of devstack.

It seems like the mass of +1s is because most developers would find this really useful. I would actually argue that for most small to medium size OpenStack deployments, this is super useful default behavior, so it would really be better if the way the neutron agents worked is defaulting to using the host resolv.conf. Only service providers are likely to actually segment DNS environments between hosts and guests. The fact that there is a config override for that already in neutron seems sufficient to handle that case.

But having an out of the box default that does the nice thing for smaller installs seems like what we should do.

Patch-set: 4
Label: Code-Review=-1