Update patch set 4
Patch Set 4: Workflow-1

So, thank you everyone for all the feedback. I think this is less about DNS and more about an ideological divide. Namely, do we try to make everything "just work" as much as possible, openstack-wide, for both devstack and production, or do we leave what can be critical configuration in the hands of the operators.  

For the former, we risk leaking data, or providing a default that causes harm to a larger operator. For the latter, we make usage/adoption of openstack harder. And at this point, ask anyone that's done a production install, quite a bit harder.

I don't have a strong leaning, beyond that we need to be leaning in one of those directions openstack-wide, which is larger than any one project.

Leaving open for a bit, but WIP, while we try to get consensus within the neutron community.

Patch-set: 4
Reviewer: Gerrit User 10980 <10980@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1