Update patch set 4
Patch Set 4:

Yes, usability is very much a balancing act, we all realized that, and as such it has an engineering cost.

All things considered, what's the usability gain we get for the cost involved to address this issue in a way makes happy both small and big operators? To me it's ~0.

Patch-set: 4