Cleanup for VPNaaS plugin
With a DevStack plugin added to the VPNaaS repo, this commit
removes the VPN logic from DevStack.

In addition, to allow grenade to check if the VPN service is
running for upgrades, modified the is_service_enabled, to allow
a devstack plugin to specify a service name with hyphens and
will convert to underscores to be able to call a custom
is_<service-name>_enabled method.

For this to pass tests, we need to have previously excluded
the q-vpn service from devstack based tests. Added dependency
on commits for that.

Change-Id: Iffa9901e24adbacb581425c4b38c7e8cff0da9e8
Closes-Bug: 1473475
Depends-On: I3a4d7ecb25668eb499677d05bdb68b919592e53d
Depends-On: Idd5524ed0b602408be5a53830981a8ab974b390c