Cleanup for VPNaaS plugin
With a DevStack plugin added to the VPNaaS repo, this commit
removes the VPN logic from DevStack.

In addition, to allow grenade to check if the VPN service is
running for upgrades, modified the is_service_enabled, to allow
a devstack plugin to specify a service name with hyphens and
will convert to underscores to be able to call a custom
is_<service-name>_enabled method.

Change-Id: Iffa9901e24adbacb581425c4b38c7e8cff0da9e8
Partial-Bug: 1473475