Update patch set 3
Patch Set 3: Code-Review-1

So I'd just like to get this together with If461d45246e05ec13b8951662d9e3ea3de823d9f and make sure we're doing the right thing.

I'm not a big tester of this path; that said bug #1280356 describes requiring the default images even when IMAGE_URLS is set.  

this seems to cover the case of where you've set a different cirros image, but not the case where you've left it out but you really need it

Can we get this so that it closes both bugs?

Patch-set: 3
Label: Code-Review=-1