Make ini-config handle file in directory without write perms
The top of the ini munging file says it is a stand-alone library, so
it seems like it should "just work" with files that you own that
happen to live in directories you don't, such as /etc.

"sed -i" will try to write out a sibling file as part of the in-place
fixup, which will fail in such a case.  This leaves us with either
workarounds or having to run the whole thing under sudo.

This change modifies ini-config to always work on a temporary file
when editing.  The test-case is updated to simulate this situation.

Change-Id: Ib7556dac9aaaf2f3c96237e0ca28ed6ae1b1b7ac