Update patch set 1
Patch Set 1:

The problem is that when you run this on an instance that happens to have an IP in the default range 10.0.0.0/24, get_default_host_ip will simply die. This is kind of o.k. when running stack.sh, where the user has a chance to provide her own local.rc in order to override this. But in the circumstances here - image_list.sh being called for example by nodepool/dib - there is no such solution.

Patch-set: 1
Label: Verified=0