Update patch set 3
Patch Set 3:

It's not trivial to get completely right.

As a suggestion perhaps we should always append to IMAGE_URLS as in change If461d45246e05ec13b8951662d9e3ea3de823d9f but then you need to remove duplicates.

Removing duplicates in the simple way (e.g. IMAGE_URLS=`echo $IMAGE_URLS | tr ',' '\n' | sort | uniq | tr '\n' ','`) is close, but actually the kicker is that I think upload_image expects the filename to be the unique key.  What do you expect to happen in this case?  Let the duplicate filename go through, which would result in the error from glance, or pick one of the two URLs?

I suppose the other thing we could do is change all of the additions to check if the filename is already in IMAGE_URLS before adding it, but that's going to get quite verbose and ugly.

Patch-set: 3