Update patch set 3
Patch Set 3:

@Bob

Is it possible we want two separate things here?  One is for when you want to *override* the default image, as seems to be the intent of this change.  In that case, you know what you're doing.

The other seems to be when you want to append another image for other reasons, but leave the default alone.

So maybe we should support both through different variables?

Patch-set: 3