Ensure images downloaded from a previous stack are fully downloaded
Currently, stack.sh does not check if an image file downloaded from
a previous stack.sh execution is fully downloaded before using it.
This patch updates the wget command to use the --timestamping
option which will re-download a file if the server modified date is
more recent than the local file as well as when the server file
size differs from the local file.

Change-Id: I343248f7d5e8bda19c93cb19092b5459b28244b0
Closes-Bug: #1403492