Update patch set 4
Patch Set 4:

I think that we will want to test on python 3 before setting python 3 as supported version in setup.cfg.
So I hope that we have a way that a project is run by python 3 regardless of setup.cfg.

For example checking whether USE_PYTHON3_{project_name} is 'True'.

Patch-set: 4