Add a $CHOWN_EXCLUDED_DIRECTORIES variable
Any folders listed in $CHOWN_EXCLUDED_DIRECTORIES will not have
chmod / chown run on them.

If you are using libvirt + NFS, this step can take quite a long time,
or if you are using libvirt + passthrough filesystems, you really
don't want this to run this.

This is due to the nfs check only checking if the root dir is
nfs. some setups have a local /opt/stack with other
directories nfs mounted into /opt/stack, and as such the
"is_nfs_directory" test allows the chmod / chown to run.

Change-Id: I5ebdaded3ffd0a5bc70c5e9ab5b18daefb358f58