Update patch set 2
Patch Set 2: Code-Review-1

It is really unclear to me what the use-case is for this?  

Why are we even doing a recursive chmod at the top of stack.sh anyway?  it makes sense to setup the directory and chmod it to us, but i don't see why we want to go touching it if it's already there?

if you're seeing problems with this, i think we should examine what we're doing and why a little closer

Patch-set: 2
Label: Code-Review=-1