Update patch set 2
Patch Set 2: Code-Review-1

Thanks Jordan, the things you find when you read the man page huh.

Also, the [ ! -f ] check probably shouldn't be there, as we want to try downloading always and let wget decide if it's newer or not.

i think this can be solved by just directly downloading with wget without the -O ... we don't transform the filename in any way, we just do "basename" on it

Patch-set: 2
Label: Code-Review=-1