Update patch set 4
Patch Set 4:

For example, in the case of ubuntu it is necessary to install 
python-mysqldb because sqlalchemy imports MySQLdb module.

But python-mysqldb doen't support python 3 and currently it seems not to be developed.

And we can find the following suggestion from comments in sqlalchemy.

https://github.com/zzzeek/sqlalchemy/blob/master/lib/sqlalchemy/dialects/mysql/mysqldb.py#L24

Should we use https://github.com/PyMySQL/mysqlclient-python instead?

Patch-set: 4